web_app/requirements.txt
# Django
Django==5.0.6
django-cors-headers==4.3.1
django-extra-fields==3.0.2

# Пакеты DRF
djangorestframework==3.15.1
django-filter==24.2
Markdown==3.6

# Django REST Framework API Key
djangorestframework-api-key==3.0.0

# Для цветного logger.
colorlog==6.8.2

# PostgreSQL
psycopg[binary]==3.1.19

# Celery
celery[redis]==5.4.0
celery_singleton==0.3.1
flower==2.0.1

pydantic==2.7.1
